backend/requirements.txt
asgiref==3.8.1
chardet==5.2.0
Django==5.0.7
pillow==10.4.0
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
reportlab==4.2.2
sqlparse==0.5.0
typing_extensions==4.12.2
